fix: add getAssociations helper function
